requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
spacy
python-dotenv
